Fix for: Bug 1543 - Potential infinite loop in AcpiDmDumpViot()
Reported by:
 Colin Ian King

+++ source/common/dmtbdump3.c
@@ -681,7 +681,7 @@ AcpiDmDumpViot (
     ACPI_TABLE_VIOT         *Viot;
     ACPI_VIOT_HEADER        *ViotHeader;
     UINT16                  Length;
-    UINT16                  Offset;
+    UINT32                  Offset;
     ACPI_DMTABLE_INFO       *InfoTable;
 
     /* Main table */
